requirements.txt
mypy
ruff
numpy
plotly-express
pandas
pandas-stubs